pyproject.toml
[tool.poetry]
name = "turntable"
version = "0.1.0"
description = "Turntable audio monitoring"
authors = ["Correl Roush <correl@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pyalsaaudio = "^0.9.0"
pydejavu = {git = "https://github.com/worldveil/dejavu.git"}
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.782"

[tool.poetry.scripts]
turntable = "turntable.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
